+++ pkg/webui/components/switch/index.js
@@ -15,6 +15,8 @@
 import React from 'react'
 import ReactSwitch from 'react-switch'
 
+import COLORS from '@ttn-lw/constants/colors'
+
 import Icon from '@ttn-lw/components/icon'
 
 const Switch = props => (
@@ -26,8 +28,8 @@ const Switch = props => (
     checkedIcon={
       <Icon icon="check" style={{ color: 'white', marginLeft: '5px', fontSize: '1rem' }} />
     }
-    onColor="#3D73FF"
-    activeBoxShadow="0 0 3px 5px #3D73FF66, inset 0 0 3px 1px #0002"
+    onColor={COLORS.C_ACTIVE_BLUE}
+    activeBoxShadow={`"0 0 3px 5px ${COLORS.C_ACTIVE_BLUE}66, inset 0 0 3px 1px #0002"`}
     height={24}
     width={44}
     data-test-id="switch"